Add support for no_std.
This change adds support for no_std by replacing the usage of `std`
with `core` and `alloc`. The thread module is only available with std
however.

An API change was needed for entropy:Error. This is because
std::error::Error was exposed and it seemed to messy to preserve
it.